test(cli): move dataset cases to a separate file
